Use a column-flex box for the footer to be at the bottom
If we used fixed, if there's any content at the bottom of the page (like
on a mobile) it will be cut off.

Using a flex-box, by setting the min-width of the body to 100vh, the
footer will have a margin-top that is the rest of the space of the body.
This forces the footer to the bottom and allows any content to be
scrolled to.